Force to unregister lifecycle observer in main thread
(1) CameraX may unregister lifecycle observer when the
onDestory of the Activity is called, or when
ProcessCameraProvider#shutdown is called.
(2) The ProcessCameraProvider#shutdown unregister the lifecycle
observer in the caller thread.
(3) If the test app is closed before shutdown, the issue
may not happen since the unregistering of the observer will be
invoked on the onDestory callback, which is in the main thread.

The issue happens when calling shutdown in the test thread and the
onDestory of the test Activity is not being called.
To make the implementation of the test flexible, this CL
post the clear phase of the lifecycle to the main thread.

Bug: 280481365
Test: ExistingActivityLifecycleTest
Change-Id: I2dfc2b69509c697b90cd3ffa3556099143707ea9