Build entriesToPop from transitionsInProgress and back stack
Instead of a maintaining a separately list of entries to pop based on
when popBackStack is called. We can instead build the list based on
transitionsInProgress and the back stack. If it is in
transitionsInProgress but not the back stack we know it has been popped.

RelNote: "Fixed an issue with Navigation in Fragments where navigating
with popUpTo and popping a fragment off the back stack without
recreating its view would cause system back to stop working."

Bug: 281726455
Fixes: 280884685
Test: Added NavControllerWithFragmentTest
Change-Id: Ieb8d4e6a18f8c223ff3193ea4348fb6065e9de94