Fix KSTypeVarianceResolver for type aliases with type parameters.
This CL fixes resolution for type aliases with type parameters, e.g.

```
typealias MyAlias<T> = Foo<@JvmSuppressWildcards T, Bar<Baz<T>>
```

This required fixing two main issues.

The first issue was that we were incorrectly resolving type aliases with
type arguments. In particular, we were replacing the type arguments
directly, which only works if the type alias and type have a 1-to-1 type
parameter to type argument. For example, `MyAlias<Qux>` would be
translated to `Foo<Qux>` rather than `Foo<Qux, Bar<Baz<Qux>>>`. Now, we
properly replace the type parameters with its corresponding type
argument.

The second issue was that we would immediately resolve type aliases when
creating a `KspType`, so the information of the type alias definition was
not avialable when resolving the type via `KSTypeVarianceResolver`. This
meant that we would lose information such as the `@JvmSuppressWildcards`
in the type the example above since the type arguments had already been
replaced. Now, we keep track of the original type alias in the `KspType`
so that we can pass it in when calling `KSTypeVarianceResolver`.

In addition, I restructured the `copy*` methods so that we can easily
copy a `KspType` without having to duplicate a method in every subtype.

Test: KspTypeNamesGoldenTest.kt
Change-Id: Iaa300bc140ac3d83db692d7c983cbb8ab3ef28d3