reset the default build version on test starts and register PA
A test failure report came in regarding a device with an API 30 sdk
hitting an unsupported sdk error. This should not occur since we support
sdks back to 26.  The only explanation is that the
Utils.BuildVersionAdapter is set the invalid version adapter. In order
to prevent that from happening the Utils.BuildVersionAdapter should be
reset after tests that manipulate it and in the setUp and tearDown.

Bug: 281821694
Test: refactoring CL. Existing unit tests still pass.
Change-Id: I626b36165654af63fbcf68eab11a4ba8fbe88fb1