Fixed mergeIntoMain in AGP 8.1
With Agp 8.0, due to b/265438201 we can only merge into the main baseline profile, profiles for variants of the same build type. For this reason, even when mergeIntoMain is true we generate the build type generation task, such as "generateReleaseBaselineProfile". This has been fixed with Agp 8.1 so mergeIntoMain correctly merges any variant of any build type that has the flag set to true.

Bug: 282816930
Test: ./gradlew :benchmark:benchmark-baseline-profile-gradle-plugin:test
Change-Id: I7c30a0b9e585b6b86cab42f036c2c040020ece5d