Handle Conditional Auth Module Version Parsing
For passkey retrieval, we may require a backwards compatible json
parsing call to the underlying auth module, which differs from the most
recent one. This is only needed for rare cases where the device hasn't
updated the GMS module, so it contains an older version of our auth
submodule. However, our JP library knows the methods of the more recent
auth submodule. Thus, while our library knows to call the new method,
the device will not have the method physically available to call.
This fix should ensure that in these rare cases, we fallback to the
older method.

Bug: 283271359
Test: Build Test
Change-Id: I23878189ac283ca352c88852832f4527a700645a