Do not fail 3A requests while we're submitting repeating requests
Currently when GraphProcessor submits pending repeating requests, it
pops the requests out of the queue, and releases the lock that protects
it before proceeding to submit the requests.

This causes an intricate race condition wher if, after clearing out the
queue, GraphProcessor.hasRepeatingRequest() is called, it will be
mistakened into thinking that there is no current repeating request nor
is there any pending request in the queue.

This CL fixes this by reusing the outer tryStartRepeatingExecutionBlock,
guaranteeing that while we're checking whether we have a repeating
request, we aren't also submitting repeating requests in the meantime.

Bug: 283797757
Test: GraphProcessorTest
Change-Id: I2098f153ce70f1b29a6a179e8b7a1ae5f5a25f5d