Avoids reflection usage in TAG messages
Our previous TAGs utilized a dynamic string that inherently
relied on reflection. This turns out to be bad for performance, so we
aim to get rid of that with this fix and just apply named strings.

This also changes the named constants that we used to key data across
the HiddenActivity so that they too do not rely on reflected strings.

Bug: 285197680
Test: Build Test
Change-Id: I5d2d1ac013eb5aeb53bd98e1a1e829ecb23eed65