Add toString() to work's Constraints class.
WorkInfo.toString() is overriden to provide, among other things, the
work's constraints. Right now, Constraints doesn't override "toString",
so we get the default implementation, which just prints the object's
address and provides no useful info. With this change, we print the
entire contents of Constraints in order to make WorkInfo.toString() more
useful.

Also fixing a typo in WorkInfo.toString() that was apparently added in
https://android-review.git.corp.google.com/c/platform/frameworks/support/+/2478188.

Test: ./gradlew :work:work-runtime:test
Change-Id: I1bbe2a18217400c6610e1f52fef92b07c13488f0