Enable benchmark-kmp darwin snapshots
This CL enables snapshots for the darwin target of benchmarks.

There are 3 notable changes:

* Added a an empty jvm target to both projects
This is needed because w/o a jvm target, no components are created
from the and publishing cannot find anything to publish.
https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:buildSrc/private/src/main/kotlin/androidx/build/MavenUploadHelper.kt;l=70
As part of that changed, moved the code from `common` to `darwinCommon`.

* guarding ios setup behind an if check to avoid "unrecognized message" errors in the linux target (it complains about configured sourcesets that are not used because ios is disabled)

* Added an empty source file to darwin-core
Without a source file, no klib is created which fails publishing.
see: https://youtrack.jetbrains.com/issue/KT-52344

Also setting JVM is the default platform. It does not make sense but
we need to pick a platform that is built on all machines so that the
maven pom file can be generated properly. This does not impact clients
that consume gradle metadata.

Test: CI
Change-Id: Ibb2421021462c3e965c8297f00aa02dee65be197