Reset Preview pipeline before activate repeating output
Preview notified the `active` state before `reset` the pipeline;
it may cause the Preview to start the repeating output on the previously
configured Surface.
Modify the call sequence, `reset` the pipeline before switching to
`active` state. This can prevent the repeating request to be started
before the new Surface to be configured.

Bug: 287140309
Test: VideoEncoderTest, PreviewViewDeviceTest, PreviewViewFragmentTest,
PreviewLifecycleStatusChangeStressTest, PreviewSwitchCameraStressTest

Change-Id: I34eee126862eb2b3210e281d7e9673ae9b60ffeb