AOT compile microbenchmarks when run locally
Fixes: 288124146
Bug: 287358254
Test: installRelease several benchmarks on API 30 & 23
Test: LazyListScrollBenchmark.scrollProgrammatically_newItemComposed

Enable full AOT for all microbenchmarks to increase stability in local
runs. This works around an issue where local runs of compose
microbenchmarks take an extremely long time to converge.

This reinstates the ancient hack:
https://android-review.googlesource.com/c/platform/frameworks/support/+/721678

That compilation hack was originally removed due to improvements to
warmup, but since that landed, the amount of non-platform code that
can be in a microbenchmark has increased drastically, to the point
where it's difficult to stabilize in a reasonable amount of time (>30s
in some cases), and difficult for warmup to detect stabilization.

Verified compilation with:

> adb shell dumpsys package dexopt | grep -A1 androidx.

Change-Id: I726d0e4ed5424c3f6e61850ca76f7576b4d94bb7