Chip and Card composables for SwipeToReveal
Adding two new composables for SwipeToReveal to be used
with Cards and Chips.

Relnote: """
We have added two new composables in material for implementing
SwipeToReveal with Cards and Chips. These composables follow
the recommended UX guidance on the component and makes it
easier for developers to implement SwipeToReveal with the
existing components in wear material.
"""
Test: TODO
Bug: 279859390

Change-Id: I7ec65ff3e6c419685f5f3e80a3073ee78143feb0