Omitting an unnecessary focus event
When we request focus for a child, first the parent
is granted focus and then focus is transferred to
the child. This results in an extra focus event for
the parent.

This CL refactors the requestFocusForChild() function
so that it does not automatically send focus events.
The focus events are sent from the code that calls
requestFocusForChild().

Bug: 275633128
Test: Added new tests to FocusEventCountTest
Change-Id: Ie338d4370f0b15d9b5df0274fe73316482790c08