Adding support for nested map return types in DAO functions of Room.
This change involves a refactoring of MapQueryResultAdapter to handle both regular maps (with or without collection value types) as well as nested map values. There is no limit set to the amount of nesting, and the algorithm is able to handle as many nested maps as possible.

Currently, @MapInfo support is limited to non-nested map return types and this functionality will be available in a follow-up CL. In addition, support for ArrayMap, LongSparseArray or IntSparseArray will be added in a future CL.

Bug: 203008711
Test: MultimapQueryTest.kt, DaoKotlinCodeGenTest.kt
Change-Id: I13f488d2d26b5f41cebf14c2cc789eb291aa064c