Update AndroidX to coroutines 1.7.1
Update MultiProcessDataStoreMultiProcessTest/DataMigrationInitializerTest use of dispatchTimeoutMs
Updated Compose async font tests to use the new dispatchers
Updated LazyPagingItemsTest to not assume context's isActive will be false on creation. In fact, this
is true as soon as it is created before it is used.
Updated GlanceAppWidgetReceiverTest to pass again. Made some changes in the code but overall also moved
it off of the coroutines testing APIs since it does not really use them. I tried to cleanup the hosts
but there is too much I don't understand to fully move those tests to coroutines.

Bug: b/285151193
Bug: b/289542327
Test: ./gradlew :datastore:datastore-core:test; ./gradlew :datastore:datastore-core:connectedAndroidTest
Change-Id: I4716cb02c83b5fa451fcb8af1f8fe0eb14958b57