Fix RxDataStore isDisposed
isDisposed was returning isActive instead of isComplete ¯\_(ツ)_/¯

I also tried to repro b/288539336. It is happening because
shutdown hook has no way of waiting for the removal of the file
lock. This test does not even care about it and fixing b/233402915
will fix that issue.

Otoh, i was able to run it 100 times on FTL w/o failure so I'm not
sure how prominant this error is. If it keeps happening, we might
put in a delay for now (very sad).

Fixes: 290668148
Bug: 288539336
Bug: 233402915
Test: RxdataStoreBuilderTest

Change-Id: I1430e481b0906710a480544b9d5955dda025f1b2