1.5.0-beta03 merge. Fix CoreText.kt (#664)
- Apply changes from jb-main between the base commit 82b269dac2cf1e4b23ec143150c920ac58e40402 and the pre-merge commit 50f7d4c072ebd19e6633031f0a94eea1a32eb0b9:

  - 20668654 (this commit was a squash, and its message is a mess)

The `pointerInput(Unit)` is restarted every recomposition, because we create `mouseSelectionObserver` every time.

It can cause selection issues. We can accidentaly not have issues (they depend on how `mouseDragGestureDetector` is used inside TextField), but it can return in the future, so better to write these modifiers in a proper way.