Updated Button, IconButton and TextButton following updates to Material3 design
Ensure all text fits on the screen with largest font sizes.
Allow semantic role to be overridden for IconButton and TextButton.
Increased the number of items in ScrollAway demos to help testing with rotary.

Bug: 291062960
Bug: 291062329
Bug: 291061710
Bug: 291060301
Bug: 291060489
Bug: 291061410
Bug: 290617178

Test: Manual testing + screenshot tests
Relnote: "We have updated the colors for Button, IconButton and TextButton in line with Material3 design. The semantic role for Button, IconButton and TextButton can now be overridden using Modifier.semantics."
Change-Id: Ib24954145bc5440092cc2ac7dd18d3bc4fd4b756