Creating the @MapColumn annotation to replace @MapInfo.
This CL adds support for the new @MapColumn annotation which declares which column(s) are used to build a map or multimap return value in a DAO query method. This annotation replaces the @MapInfo annotation and should be used on the type specifically. This CL also deprecates the @MapInfo annotation.

In a follow-up CL, support for LongSparseArray, ArrayMap and IntSparseArray will be added for MapColumn and nested maps return types.

Bug: 203008711
Test: QueryMethodProcessorTest.kt, RawQueryMethodProcessorTest.kt, MultimapQueryTest.kt
Relnote: Creating the `@MapColumn` annotation to replace `@MapInfo`, which is now deprecated.
Change-Id: Ib0305c9ba7c0bd73fd885f22b5fac4f6c90f0b24