Disable element validation in Hilt Worker extension.
In the conversion to support KSP (aosp/2658055) element deferring was enabled since it is the default behavior in `XBasicAnnotationProcessor` but this can lead to issues with the `RootProcessor` missing generated modules due to the detection mechanism checking the round environment. This changes disable deferring and validation keeping old behavior until a solution is found for `RootProcessor`.

Bug: 292261505
Test: n/a
Change-Id: Ie521d70ec169e2b5112bdc117fa12f9efb65a6fb