Fixed the extensions native crash caused by camera3_jpeg_blob data missing
When capturing photos using Extensions, we invoke the native codes
nativeWriteJpegToSurface to write the jpeg bytes into the Surface.

However, the native layer expects jpeg bytes to contain the camera3_jpeg_blob
struct at the end of the buffer. If jpeg bytes are supplied without the
camera3_jpeg_blob, it is possible that the content byte matches the
CAMERA3_JPEG_BLOB_ID by chance and cause the wrong jpeg size to be reported.

To workaround the problem, here it adds the padding 0s to the end of the buffer
so that CAMERA3_JPEG_BLOB_ID won't be matched by any chance and the total bytes
size is reported as the jpeg size accordingly.

The side effect of this approach is that there will be 8 zero bytes at the end
of the jpeg bytes apps received. Because the camera3_jpeg_blog is private definition
in native layer and it is changed to camera3_jpeg_blog_v2 in Android 13, it is
not suitable to write the struct in CameraX.

Bug: 289021003
Test: ImageProcessingUtilTest
Relnote: "Fixed the native crash when taking pictures with Extensions enabled"
Change-Id: I698f5739cd3d42243721184c1162f0ba86523e0d