Migrate JvmDescriptorUtils to utilize Elements.getBinaryName()
This allows the utils to support anonymous classes that are unnamed. It is not possible to discover a ANONYMOUS type element compiled from Java sources via javac, but kotlinc does in some occasions produces bytecode that uses a ANONYMOUS type in a method signature. Specifically for private lazy delegate properties with an anonymous expression and whose type is inferred.

The need to use descriptors for such types is important for performing Kotlin metadata lookup, for example, to identify the synthetic getter method for a lazy delegate property, xprocessing needs to be able to match descriptors even if the getter returns an anonymous type.

Moreover, this change also fixes the inconsistency of lazy delegate backing field being discoverable in javac / KAPT but not KSP, the decision being making them invisible for javac / KAPT since whether a delegate creates a backing field or not is an implementation detail of the compiler.

Fixes https://github.com/google/dagger/issues/3962
Test: KotlinMetadataElementTest

Change-Id: Id8dfe0b21b29f895ef97fdcb89a00ea91a660cc0