Make sure the time update callback is always called after init
When there are two registrations for platform time data source, the
second one only receives a result through the scheduled tick (not with
the initial registration). This change moves the initial tick call
from the notifier to the caller. This way we can wait for all
registrations and then send the callback once.

Relnote: "Fixed an issue when an expression with with mutliple time
data source registrations was not being updated immediately."
Bug:292071881

Change-Id: I8e1a80ea606a831433f81c9454605803a24b4c8e