Bugfix: close parent SurfaceEdge's DeferrableSurface when the child is closed.
In a stream sharing scenario, StreamSharing is the parent and Preview and VideoCapture are the children. When a child's DeferrableSurface is closed, e.g. by SurfaceViewImplementation#surfaceDestroyed(), it indicates that the app wishes to stop receving frames on that Surface. We need to notify the OpenGL renderer in StreamSharing to stop drawing. Currently, this is missing.

In this CL, we allow the SurfaceEdge to listen to the provider DeferrableSurface's closure, and disconnect when it closes. Disconnecting the edge prevents the OpenGL from drawing to the Surface

- Add a #getCloseFuture() method to DeferrableSurface. This future finishes when the DeferrableSurface is closed.
- In SurfaceEdge, listen for the provider's close future and disconnect. Disonnection will notify the SurfaceProcessor to stop drawing.
- In DefaultSurfaceProcessor, catch and log the crash just to be safe.

Note that the error may still happen even after this fix, because SurfaceView can release the Surface faster than OpenGL can stop rendering. However this should only happen to older devices with a faulty GL driver. I have not been able to reproduce this locally. But if it happens, we need to force those devices to use TextureView.

For details see: go/camerax-image-stream

Bug: 294521827
Test: manual test and ./gradlew bOS
Change-Id: I884e4dee5abc1d44c3dfa0c329ed7bca05b55179