Keep invalidated ComplicationData when keepDynamicValues is true.
This allows using the dynamic values even when the evaluation was
invalidated, using the new NoDataComplicationData.invalidatedData field.

More context:

When ComplicationData is evaluated, the dynamic values either form a
valid result or an invalid result. The main reason for an invalid result
is if the data is not yet available, i.e. between the time we registered
for the data (e.g. in WHS) and the first callback for that registration.

If the ComplicationData evaluates to a valid value,
keepDynamicValues=true mean that the evaluator should keep the original
unevaluated dynamic values (aka expressions) alongside the evaluated
ones, in their original fields (e.g. ComplicationText will have both
surroundingText and dynamicText set).

If the ComplicationData evaluates to an invalid value, the evaluator
instead returns NoDataComplicationData with
placeholder=original.placeholder. This means that keepDynamicValues=true
is ignored, as the entire original ComplicationData is trashed (only the
placeholder field is used from it).

The problem with this is described in the associated bug, and solved by
providing the original data in a separate field.

Bug: 294542147
Change-Id: Ie7d821121d3c3c0332f9f8d6c1942a1fbed07145