Add new API to fix TestLifecycleOwner hanging inside of runTest
The use of runBlocking in TestLifecycleOwner causes
tests to hang when the api is used from the context of
another coroutine (such as runTest). To fix this, we
have added a suspending function to allow using
TestLifecycleOwner from within a coroutine.

RelNote: "`TestLifecycleOwner` now includes the suspending
function `setCurrentState()` to give users the option of using `TestLifecycleOwner` from within a coroutine."
Test: testSetCurrentStateInRunTest
Fixes: 259344129

Change-Id: I329de3d76bcd7fa5e04077916ff551f9f2f3f864