Create Fallback Velocity Tracking for AXIS_SCROLL
This change adds a fallback AXIS_SCROLL velocity tracking logic, to
support the axis' velocity tracking for pre-Android-U versions.

To be able to fully offload the velocity tracking work for this AXIS to
the fallback tracker, we also add compat APIs for some VelocityTracker
APIs.

RelNote: adds more compat APIs for Velocity Tracker
Bug: 294895104
Test: manual, unit tests
Change-Id: I327530551bbae8e4594d1b081ce3277bc60efe57