Fix an issue with popping the last back stack entry
When using predictive back if you pop the last entry on the back stack,
attempting to go back to the first fragment, the app will skip the first
fragment and then destroy the activity.

This happens because when predictive back starts a back gesture, the
transition in progress is no longer includes as part of the back stack
count, which cause the fragment manager to prematurely disabled its
onBackCallback thinking that it has no entries and then it finishes the
activity instead of going back.

We should ensure we continue to consider the transitioning operations as
part of the back stack while transitioning.

RelNote: "Fixed an issue with predictive back that causes the Activity
to be finished, before ever seeing the first fragment in the fragment
manager."
Test: added PredictiveBackTest
Bug: 295231788
Fixes: 295168369

Change-Id: I0664b44dd7f6dc49bafb932e3ae47f1aba483fd6