Make sure to finalize sessions when session configuration fails
Due to timing differences, capture session configuration can happen as
the camera device is being closed. Under this scenario, an exception is
thrown when calling createCaptureSession(), but onConfigureFailed()
on the other hand, would not be invoked on the callback we called it
with.

This CL makes some adjustments such that when we fail to create a
capture session, we always finalize the current, as well as the previous
capture session.

Bug: 295125234
Test: Run ToggleButtonUITest 25 times consecutively.
Change-Id: Ib40bba0356ab3336a13e2b7049cabde3d6a9fbe8