Rework synchronization for concurrent cameras
This CL is a partial rework and refactoring of the synchronization
mechanism originally created for concurrent cameras.

Whenever we have use cases to bind to the lifecycle, we have the
following procedure:
1. CameraCoordinator.setCameraOperatingMode()
2. ProcessCameraProvider.bindToLifecycle()
3. CameraCoordinator.setActiveConcurrentCameraInfos(), if under
   concurrent camera mode.

Currently, we basically treat setCameraOperatingMode() as a signal for
whether to proceed to creating a CameraGraph. When we're under
concurrent camera mode, we hold off on any CameraGraph creations until
setActiveConcurrentCameraInfos() is invoked.

Now the tricky bit is once we invoke setActiveConcurrentCameraInfos(),
we need to wait till all CameraGraph configs are received before
creating a list of CameraGraphs at CameraPipe.

CameraGraphCreator was created for this, and it basically blocks any
CameraGraph creation callers until it receives all CameraGraph.Configs.
This means we had to make sure we create CameraGraphs on different
(background) threads.

The side effect of that, unfortunately, is that
CameraGraph creations might be delayed and more importantly out of sync
with the UseCase lifetimes maintained at CameraInternalAdapter and
UseCaseManager.

To handle this, on a high level, we need to make sure we always create
CameraGraphs on the calling thread (main thread) when use cases are
binded to the lifecycle. How we accomplish this is,
1. We first allow UseCaseManagers to create a "deferred"
   UseCaseManagerConfig that contains a CameraGraph.Config and all
   necessary objects to create the UseCaseCamera components.
2. If the camera operating mode is single, we create the components
   right away. If it is concurrent, we store the "deferred"
   UseCaseManagerConfig.
3. When setActiveConcurrentCameraInfos(), we gather all the
   CameraGraph.Configs from each of the UseCaseManager, and create the
   list of CameraGraphs at CameraPipe.
4. We then proceed to "resume" the previously deferred
   UseCaseManagerConfigs as well as provide its corresponding
   CameraGraph.

Bug: 295805791
Test: ConcurrentCameraActivity and coretestapp
Change-Id: Ic3288147cb7e0c4f9300b8466fb27b74bb38e65f