Introduce HapticFallbackChainSignal
Introduce HapticSignal extensions to allow custom definitions of haptic
signals that can be resolved by the HapticManager to be played by a
device vibrator. This introduces the types:

- ResolvableHapticSignal
- HapticFallbackChainSignal

This change also introduces the device support check to
HapticDeviceProfile, which can be used by the fallback chain
implementation to resolve the first fully supported HapticSignal.

Samples and demos are also introduced by this change.

Bug: 187296326
Bug: 275083960
Test: unit tests in androidx.core.haptics
Relnote: "Introduce HapticFallbackChainSignal to represent
basic fallback strategy based on device SDK and hardware support"

Change-Id: I9e1b00c36fb62813bd741e700005adfc07cf605b