Deprecates NotificationCompatSideChannelService
Adds a DeprecatedSinceApi(19) annotation to
NotificationCompatSideChannelService.java. Side channeling shouldn't be
used beyond API level 18, but this API was accidentally made available
in 19 as well. Adding this annotation so that when the min API level
increases again, we'll be reminded to clean this up.

The current class doesn't actually do anything for APIs beyond 19
anyway.

Bug: 296630027
Test: Adding annotation, no code change
Relnote: Marks NotificationCompatSideChannelService as deprecated
Change-Id: I18fd939b593bd4b624d5c8e09358b8cb3893da55