Revert "Revert "Instrument tests for glRenderer""
This reverts commit 4e417369a7f0ddb9d52aa51576fd6b655f15e8bf.

Reason for revert: Fix test and roll forward

The failure is caused by the GL thread check. We can't reproduce this locally or with ACID. In this CL, we post all the gl operations on an executor and see if it makes the issue go away. The fix is not verified because we can't reproduce the failure.

Bug: 295407763
Test: manual test and tested with ACID device
Change-Id: I85ed7b9d49a2feef118460dfc2f864e2947a1a83