Add support for dropping java resources in BundleInsideHelper
This allows moving the dropping of *.proto files to the projects
that bundle the libraries. In turn, this means we no longer need
to create custom configurations in the included projects when referencing
them in the including projects.

This is a step towards making it easier to resolve what artifact
coordinates were actually bundled inside.

Test: ./gradlew appactions:interaction:interaction-service:publish
      comparing ../../out/androidx/build/support_repo/androidx/appactions/interaction/interaction-service/1.0.0-alpha01/interaction-service-1.0.0-alpha01.aar
      before and after the change, libs/repackaged.jar is identical
Bug: 296864329
Change-Id: If749f1ea0dce519be9d5b11322477227e6aa7eec