Reduce number of a11y events sent during scrolling
This moves the VIEW_SCROLLED and CONTENT_CHANGED events to get
batched with the usual 100ms delay instead of being sent immediately on
each scroll frame, and also ensures only one of each is sent per
scroller.  I verified this matches the frequency of events sent by
ScrollView.

The key trick in this patch is to keep holding a reference to the
AccessibilityNodeInfo that ViewRootImpl uses to store the bounds of the
green focus box. We then cheaply update its bounds in place, instead of
using a CONTENT_CHANGED event to trigger its recreation. This trick is
necessary to match the performance of View system, because it replicates
how ViewRootImpl.getAccessibilityFocusedRect examines the live View
hierarchy if `provider == null`.

Bug: 191820429
Bug: 291662443
Test: sendScrollEvent_byStateObservation

Change-Id: I33a6238dfc9fa8e870da46e0a2a1ace4bca651ed