Clear and freeze active complication data on edit.
This is to avoid momentarily seeing the old complication after editing.

It's not perfect - there can be a race where the platform sends new data
from the old source between unfreezing and the platform switching to the
new source, especially in high frequency complications.

Bug: 194776118
Tested: Change to different complication, same complication (restore in the same edit session), different configuration.
Change-Id: Idb379d24612b933efd7b098e9d156bb677bd54cf