Fix DialogFragment not marked transition complete
When navigating to new DialogFragment, the current open DialogFragment (outoing entry) is marked as transitioning. Likewise, when popping a DialogFragment, the incoming entry is marked as transitioning. However in both cases, the entry is not marked complete after navigation/pop. In the second case, it prevents the incoming entry from moving to RESUMED state.

Now we ensure to remove the entries from its transient transitioning state so it can move to proper lifecycle state.

Test: ./gradlew navigation:navigation-fragment:cC
Bug: 287505132
Change-Id: I3b866a44291e7bd7af742562ae0619cd83929a1a